svn merge -r137:151 https://svn.library.emory.edu/svn/openemory/branches/release-0.2.x
various fixes/patches from staging

+++ openemory/harvest/static/js/queue.js
@@ -10,52 +10,62 @@ $(document).ready(function () {
     // configure ingest buttons to POST pmcid when clicked
     $(".ingest").click(function() {
         var data = { pmcid: $(this).find(".pmcid").html()};
-        var msg = $(this).parent().find('.message');
+        var entry = $(this).parent();
+        var msg = entry.find('.message');
         var errclass = 'error-msg';
         var success_class = 'success-msg';
-        $.ajax({
-            type: 'POST',
-            data: data,
-            url: $(this).attr('href'),
-            headers: {'X-CSRFTOKEN': csrftoken},
-            success: function(data, status, xhr) {
-                msg.html(data || 'Success');
-                msg.removeClass(errclass).addClass(success_class).show().delay(1500).fadeOut();
-                // change queue item class on success so display can be updated
-                msg.parents('.queue-item').addClass('ingested');
-            },
-            error: function(data, status, xhr) {
-                msg.html('Error: ' + data.responseText);
-                msg.removeClass(success_class).addClass(errclass).show();
-                // NOTE: not fading error message out, since it may
-                // need to be reported
-            }
 
-        });
+        if ( ! entry.hasClass('working') ) {
+          entry.addClass('working');
+          $.ajax({
+              type: 'POST',
+              data: data,
+              url: $(this).attr('href'),
+              headers: {'X-CSRFTOKEN': csrftoken},
+              success: function(data, status, xhr) {
+                  msg.html(data || 'Success');
+                  msg.removeClass(errclass).addClass(success_class).show().delay(1500).fadeOut();
+                  // change queue item class on success so display can be updated
+                  entry.removeClass('working').addClass('ingested');
+              },
+              error: function(data, status, xhr) {
+                  msg.html('Error: ' + data.responseText);
+                  msg.removeClass(success_class).addClass(errclass).show();
+                  // NOTE: not fading error message out, since it may
+                  // need to be reported
+              }
+
+          });
+        }
         return false;
     });
 
     // configure ignore buttons to DELETE specified url when clicked
     $(".ignore").click(function() {
-        var msg = $(this).parent().find('.message');
+        var entry = $(this).parent();
+        var msg = entry.find('.message');
         var errclass = 'error-msg';
         var success_class = 'success-msg';
-        $.ajax({
-            type: 'DELETE',
-            url: $(this).attr('href'),
-            headers: {'X-CSRFTOKEN': csrftoken},
-            success: function(data, status, xhr) {
-                msg.html(data || 'Success');
-                msg.removeClass(errclass).addClass(success_class).show().delay(1500).fadeOut();
-                // change queue item on success so display can be updated
-                msg.parents('.queue-item').addClass('ignored');
-            },
-            error: function(data, status, xhr) {
-                msg.html('Error: ' + data.responseText);
-                msg.removeClass(success_class).addClass(errclass).show();
-            }
 
-        });
+        if ( ! entry.hasClass('working') ) {
+          entry.addClass('working');
+          $.ajax({
+              type: 'DELETE',
+              url: $(this).attr('href'),
+              headers: {'X-CSRFTOKEN': csrftoken},
+              success: function(data, status, xhr) {
+                  msg.html(data || 'Success');
+                  msg.removeClass(errclass).addClass(success_class).show().delay(1500).fadeOut();
+                  // change queue item on success so display can be updated
+                  entry.removeClass('working').addClass('ignored');
+              },
+              error: function(data, status, xhr) {
+                  msg.html('Error: ' + data.responseText);
+                  msg.removeClass(success_class).addClass(errclass).show();
+              }
+
+          });
+        }
         return false;
     });
